Add sigint capture functionality to show warning
Changes are first saved in a dict, which is then saved back to a data
file. Data integrity cannot be guaranteed if the program was closed while
running. It takes some time to save the data from memory back to disk.
If the program is stopped while writing the file, the resulting file
will be incomplete and corrupted. The more data, the longer it takes to
save it all back into a file.
If the data is saved to disk less often, the probability of corrupting
data is lower. However in that case, the information in the data and
the information online might differ. If the program is closed after
creating entries online, but before saving those entries to the local
disk, there will be a difference between the data instead of an error.
With the current implementation there is a tradeoff between data
integrity and data sync.
TODO: Fix this mechanism to guarantee data integrity while keeping both
online and offline versions  in sync. Maybe by adding a special tag that
can be read to recognize a bug as created by this program.